Splitting remaining common class, moving "common" class to ConnectionString namespace and renaming to sync class/file name.
