+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/Common/ConnectionString/DbConnectionStringUtilities.cs
@@ -11,7 +11,7 @@ using Microsoft.Data.SqlClient;
 
 namespace Microsoft.Data.Common
 {
-    internal static class DbConnectionStringBuilderUtil
+    internal static class DbConnectionStringUtilities
     {
         internal static bool ConvertToBoolean(object value)
         {
@@ -102,131 +102,6 @@ namespace Microsoft.Data.Common
                 throw ADP.ConvertFailed(value.GetType(), typeof(string), e);
             }
         }
-
-        #region <<PoolBlockingPeriod Utility>>
-        internal static bool TryConvertToPoolBlockingPeriod(string value, out PoolBlockingPeriod result)
-        {
-            Debug.Assert(Enum.GetNames(typeof(PoolBlockingPeriod)).Length == 3, "PoolBlockingPeriod enum has changed, update needed");
-            Debug.Assert(value != null, "TryConvertToPoolBlockingPeriod(null,...)");
-
-            if (StringComparer.OrdinalIgnoreCase.Equals(value, nameof(PoolBlockingPeriod.Auto)))
-            {
-                result = PoolBlockingPeriod.Auto;
-                return true;
-            }
-            else if (StringComparer.OrdinalIgnoreCase.Equals(value, nameof(PoolBlockingPeriod.AlwaysBlock)))
-            {
-                result = PoolBlockingPeriod.AlwaysBlock;
-                return true;
-            }
-            else if (StringComparer.OrdinalIgnoreCase.Equals(value, nameof(PoolBlockingPeriod.NeverBlock)))
-            {
-                result = PoolBlockingPeriod.NeverBlock;
-                return true;
-            }
-            else
-            {
-                result = DbConnectionStringDefaults.PoolBlockingPeriod;
-                return false;
-            }
-        }
-
-        internal static bool IsValidPoolBlockingPeriodValue(PoolBlockingPeriod value)
-        {
-            Debug.Assert(Enum.GetNames(typeof(PoolBlockingPeriod)).Length == 3, "PoolBlockingPeriod enum has changed, update needed");
-            return value == PoolBlockingPeriod.Auto || value == PoolBlockingPeriod.AlwaysBlock || value == PoolBlockingPeriod.NeverBlock;
-        }
-
-        internal static string PoolBlockingPeriodToString(PoolBlockingPeriod value)
-        {
-            Debug.Assert(IsValidPoolBlockingPeriodValue(value));
-
-            return value switch
-            {
-                PoolBlockingPeriod.AlwaysBlock => nameof(PoolBlockingPeriod.AlwaysBlock),
-                PoolBlockingPeriod.NeverBlock => nameof(PoolBlockingPeriod.NeverBlock),
-                _ => nameof(PoolBlockingPeriod.Auto),
-            };
-        }
-
-        /// <summary>
-        /// This method attempts to convert the given value to a PoolBlockingPeriod enum. The algorithm is:
-        /// * if the value is from type string, it will be matched against PoolBlockingPeriod enum names only, using ordinal, case-insensitive comparer
-        /// * if the value is from type PoolBlockingPeriod, it will be used as is
-        /// * if the value is from integral type (SByte, Int16, Int32, Int64, Byte, UInt16, UInt32, or UInt64), it will be converted to enum
-        /// * if the value is another enum or any other type, it will be blocked with an appropriate ArgumentException
-        ///
-        /// in any case above, if the converted value is out of valid range, the method raises ArgumentOutOfRangeException.
-        /// </summary>
-        /// <returns>PoolBlockingPeriod value in the valid range</returns>
-        internal static PoolBlockingPeriod ConvertToPoolBlockingPeriod(string keyword, object value)
-        {
-            Debug.Assert(value != null, "ConvertToPoolBlockingPeriod(null)");
-            if (value is string sValue)
-            {
-                // We could use Enum.TryParse<PoolBlockingPeriod> here, but it accepts value combinations like
-                // "ReadOnly, ReadWrite" which are unwelcome here
-                // Also, Enum.TryParse is 100x slower than plain StringComparer.OrdinalIgnoreCase.Equals method.
-
-                if (TryConvertToPoolBlockingPeriod(sValue, out PoolBlockingPeriod result))
-                {
-                    return result;
-                }
-
-                // try again after remove leading & trailing whitespaces.
-                sValue = sValue.Trim();
-                if (TryConvertToPoolBlockingPeriod(sValue, out result))
-                {
-                    return result;
-                }
-
-                // string values must be valid
-                throw ADP.InvalidConnectionOptionValue(keyword);
-            }
-            else
-            {
-                // the value is not string, try other options
-                PoolBlockingPeriod eValue;
-
-                if (value is PoolBlockingPeriod period)
-                {
-                    // quick path for the most common case
-                    eValue = period;
-                }
-                else if (value.GetType().IsEnum)
-                {
-                    // explicitly block scenarios in which user tries to use wrong enum types, like:
-                    // builder["PoolBlockingPeriod"] = EnvironmentVariableTarget.Process;
-                    // workaround: explicitly cast non-PoolBlockingPeriod enums to int
-                    throw ADP.ConvertFailed(value.GetType(), typeof(PoolBlockingPeriod), null);
-                }
-                else
-                {
-                    try
-                    {
-                        // Enum.ToObject allows only integral and enum values (enums are blocked above), raising ArgumentException for the rest
-                        eValue = (PoolBlockingPeriod)Enum.ToObject(typeof(PoolBlockingPeriod), value);
-                    }
-                    catch (ArgumentException e)
-                    {
-                        // to be consistent with the messages we send in case of wrong type usage, replace
-                        // the error with our exception, and keep the original one as inner one for troubleshooting
-                        throw ADP.ConvertFailed(value.GetType(), typeof(PoolBlockingPeriod), e);
-                    }
-                }
-
-                // ensure value is in valid range
-                if (IsValidPoolBlockingPeriodValue(eValue))
-                {
-                    return eValue;
-                }
-                else
-                {
-                    throw ADP.InvalidEnumerationValue(typeof(ApplicationIntent), (int)eValue);
-                }
-            }
-        }
-        #endregion
 
         internal static bool TryConvertToApplicationIntent(string value, out ApplicationIntent result)
         {
@@ -707,248 +582,6 @@ namespace Microsoft.Data.Common
                 }
             }
         }
-
-        #region <<AttestationProtocol Utility>>
-        /// <summary>
-        ///  Convert a string value to the corresponding SqlConnectionAttestationProtocol
-        /// </summary>
-        /// <param name="value"></param>
-        /// <param name="result"></param>
-        /// <returns></returns>
-        internal static bool TryConvertToAttestationProtocol(string value, out SqlConnectionAttestationProtocol result)
-        {
-            if (StringComparer.InvariantCultureIgnoreCase.Equals(value, nameof(SqlConnectionAttestationProtocol.HGS)))
-            {
-                result = SqlConnectionAttestationProtocol.HGS;
-                return true;
-            }
-            else if (StringComparer.InvariantCultureIgnoreCase.Equals(value, nameof(SqlConnectionAttestationProtocol.AAS)))
-            {
-                result = SqlConnectionAttestationProtocol.AAS;
-                return true;
-            }
-            else if (StringComparer.InvariantCultureIgnoreCase.Equals(value, nameof(SqlConnectionAttestationProtocol.None)))
-            {
-                result = SqlConnectionAttestationProtocol.None;
-                return true;
-            }
-            else
-            {
-                result = DbConnectionStringDefaults.AttestationProtocol;
-                return false;
-            }
-        }
-
-        internal static bool IsValidAttestationProtocol(SqlConnectionAttestationProtocol value)
-        {
-            Debug.Assert(Enum.GetNames(typeof(SqlConnectionAttestationProtocol)).Length == 4, "SqlConnectionAttestationProtocol enum has changed, update needed");
-            return value == SqlConnectionAttestationProtocol.NotSpecified
-                || value == SqlConnectionAttestationProtocol.HGS
-                || value == SqlConnectionAttestationProtocol.AAS
-                || value == SqlConnectionAttestationProtocol.None;
-        }
-
-        internal static string AttestationProtocolToString(SqlConnectionAttestationProtocol value)
-        {
-            Debug.Assert(IsValidAttestationProtocol(value), "value is not a valid attestation protocol");
-
-            return value switch
-            {
-                SqlConnectionAttestationProtocol.AAS => nameof(SqlConnectionAttestationProtocol.AAS),
-                SqlConnectionAttestationProtocol.HGS => nameof(SqlConnectionAttestationProtocol.HGS),
-                SqlConnectionAttestationProtocol.None => nameof(SqlConnectionAttestationProtocol.None),
-                _ => null
-            };
-        }
-
-        internal static SqlConnectionAttestationProtocol ConvertToAttestationProtocol(string keyword, object value)
-        {
-            if (value == null)
-            {
-                return DbConnectionStringDefaults.AttestationProtocol;
-            }
-
-            if (value is string sValue)
-            {
-                // try again after remove leading & trailing whitespaces.
-                sValue = sValue.Trim();
-                if (TryConvertToAttestationProtocol(sValue, out SqlConnectionAttestationProtocol result))
-                {
-                    return result;
-                }
-
-                // string values must be valid
-                throw ADP.InvalidConnectionOptionValue(keyword);
-            }
-            else
-            {
-                // the value is not string, try other options
-                SqlConnectionAttestationProtocol eValue;
-
-                if (value is SqlConnectionAttestationProtocol protocol)
-                {
-                    eValue = protocol;
-                }
-                else if (value.GetType().IsEnum)
-                {
-                    // explicitly block scenarios in which user tries to use wrong enum types, like:
-                    // builder["SqlConnectionAttestationProtocol"] = EnvironmentVariableTarget.Process;
-                    // workaround: explicitly cast non-SqlConnectionAttestationProtocol enums to int
-                    throw ADP.ConvertFailed(value.GetType(), typeof(SqlConnectionAttestationProtocol), null);
-                }
-                else
-                {
-                    try
-                    {
-                        // Enum.ToObject allows only integral and enum values (enums are blocked above), raising ArgumentException for the rest
-                        eValue = (SqlConnectionAttestationProtocol)Enum.ToObject(typeof(SqlConnectionAttestationProtocol), value);
-                    }
-                    catch (ArgumentException e)
-                    {
-                        // to be consistent with the messages we send in case of wrong type usage, replace
-                        // the error with our exception, and keep the original one as inner one for troubleshooting
-                        throw ADP.ConvertFailed(value.GetType(), typeof(SqlConnectionAttestationProtocol), e);
-                    }
-                }
-
-                if (IsValidAttestationProtocol(eValue))
-                {
-                    return eValue;
-                }
-                else
-                {
-                    throw ADP.InvalidEnumerationValue(typeof(SqlConnectionAttestationProtocol), (int)eValue);
-                }
-            }
-        }
-
-        internal static SqlConnectionEncryptOption ConvertToSqlConnectionEncryptOption(string keyword, object value)
-        {
-            if (value is null)
-            {
-                return DbConnectionStringDefaults.Encrypt;
-            }
-            else if(value is SqlConnectionEncryptOption eValue)
-            {
-                return eValue;
-            }
-            else if (value is string sValue)
-            {
-                return SqlConnectionEncryptOption.Parse(sValue);
-            }
-            else if(value is bool bValue)
-            {
-                return SqlConnectionEncryptOption.Parse(bValue);
-            }
-
-            throw ADP.InvalidConnectionOptionValue(keyword);
-        }
-
-        #endregion
-
-        #region <<IPAddressPreference Utility>>
-        /// <summary>
-        /// IP Address Preference.
-        /// </summary>
-        private readonly static Dictionary<string, SqlConnectionIPAddressPreference> s_preferenceNames = new(StringComparer.InvariantCultureIgnoreCase);
-
-        static DbConnectionStringBuilderUtil()
-        {
-            foreach (SqlConnectionIPAddressPreference item in Enum.GetValues(typeof(SqlConnectionIPAddressPreference)))
-            {
-                s_preferenceNames.Add(item.ToString(), item);
-            }
-        }
-
-        /// <summary>
-        ///  Convert a string value to the corresponding IPAddressPreference.
-        /// </summary>
-        /// <param name="value">The string representation of the enumeration name to convert.</param>
-        /// <param name="result">When this method returns, `result` contains an object of type `SqlConnectionIPAddressPreference` whose value is represented by `value` if the operation succeeds. 
-        /// If the parse operation fails, `result` contains the default value of the `SqlConnectionIPAddressPreference` type.</param>
-        /// <returns>`true` if the value parameter was converted successfully; otherwise, `false`.</returns>
-        internal static bool TryConvertToIPAddressPreference(string value, out SqlConnectionIPAddressPreference result)
-        {
-            if (!s_preferenceNames.TryGetValue(value, out result))
-            {
-                result = DbConnectionStringDefaults.IPAddressPreference;
-                return false;
-            }
-            return true;
-        }
-
-        /// <summary>
-        /// Verifies if the `value` is defined in the expected Enum.
-        /// </summary>
-        internal static bool IsValidIPAddressPreference(SqlConnectionIPAddressPreference value)
-            => value == SqlConnectionIPAddressPreference.IPv4First
-                || value == SqlConnectionIPAddressPreference.IPv6First
-                || value == SqlConnectionIPAddressPreference.UsePlatformDefault;
-
-        internal static string IPAddressPreferenceToString(SqlConnectionIPAddressPreference value)
-            => Enum.GetName(typeof(SqlConnectionIPAddressPreference), value);
-
-        internal static SqlConnectionIPAddressPreference ConvertToIPAddressPreference(string keyword, object value)
-        {
-            if (value is null)
-            {
-                return DbConnectionStringDefaults.IPAddressPreference;  // IPv4First
-            }
-
-            if (value is string sValue)
-            {
-                // try again after remove leading & trailing whitespaces.
-                sValue = sValue.Trim();
-                if (TryConvertToIPAddressPreference(sValue, out SqlConnectionIPAddressPreference result))
-                {
-                    return result;
-                }
-
-                // string values must be valid
-                throw ADP.InvalidConnectionOptionValue(keyword);
-            }
-            else
-            {
-                // the value is not string, try other options
-                SqlConnectionIPAddressPreference eValue;
-
-                if (value is SqlConnectionIPAddressPreference preference)
-                {
-                    eValue = preference;
-                }
-                else if (value.GetType().IsEnum)
-                {
-                    // explicitly block scenarios in which user tries to use wrong enum types, like:
-                    // builder["SqlConnectionIPAddressPreference"] = EnvironmentVariableTarget.Process;
-                    // workaround: explicitly cast non-SqlConnectionIPAddressPreference enums to int
-                    throw ADP.ConvertFailed(value.GetType(), typeof(SqlConnectionIPAddressPreference), null);
-                }
-                else
-                {
-                    try
-                    {
-                        // Enum.ToObject allows only integral and enum values (enums are blocked above), raising ArgumentException for the rest
-                        eValue = (SqlConnectionIPAddressPreference)Enum.ToObject(typeof(SqlConnectionIPAddressPreference), value);
-                    }
-                    catch (ArgumentException e)
-                    {
-                        // to be consistent with the messages we send in case of wrong type usage, replace
-                        // the error with our exception, and keep the original one as inner one for troubleshooting
-                        throw ADP.ConvertFailed(value.GetType(), typeof(SqlConnectionIPAddressPreference), e);
-                    }
-                }
-
-                if (IsValidIPAddressPreference(eValue))
-                {
-                    return eValue;
-                }
-                else
-                {
-                    throw ADP.InvalidEnumerationValue(typeof(SqlConnectionIPAddressPreference), (int)eValue);
-                }
-            }
-        }
-        #endregion
     }
